requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
pandas==2.2.3
numpy==2.1.0
scikit-learn==1.5.2
sentence-transformers
torch
transformers
python-dotenv==1.0.1
sqlalchemy==2.0.34
alembic
pydantic==2.9.2
Pillow==10.4.0
faiss-cpu
PyMuPDF==1.24.10
google-generativeai==0.8.3
pytesseract==0.3.10
python-multipart==0.0.9
psycopg2-binary==2.9.9
jinja2==3.1.4
opencv-python==4.10.0.84
requests==2.32.3
beautifulsoup4==4.12.3
lxml==4.9.4
tqdm==4.66.5
